provide minimum version `Import::Into` dependency
When trying to install `Data::Fake`, I observed `t/api.t` failed every test. This was because I had `Import::Into` version 1.002001 already installed, which satisfied the `0` requirement but did not support automatic loading of modules, upon which the test relies. This feature was added in version 1.002003 according to the changelog here: https://metacpan.org/changes/distribution/Import-Into

I've bumped the requirement to the latest available as of today, 1.002005.